+++ CHANGELOG.md
@@ -18,6 +18,7 @@ develop
 ### Fixes
 
 - Fix an issue where the alert about the conflict with IANA charset name was not displayed.
+- Fix an issue where the application could crash on large text change.
 - Fix an issue where the application crashed when try to save a document with Non-lossy ASCII encoding.
 - Fix an issue where MarsEdit via the App Store didn't update its contents after closing the document in CotEditor.