+++ CHANGELOG.md
@@ -23,6 +23,7 @@ develop
 - Fix an issue where some kind of files could not be opened via Service.
 - Fix an issue where the application crashed when try to save a document with Non-lossy ASCII encoding.
 - Fix an issue where MarsEdit via the App Store didn't update its contents after closing the document in CotEditor.
+- Fix syntax highlight of quoted text of which quotation delimiter consists of multiple characters.