+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 Change Log
 ==========================
 
-develop
+3.0.5 (170)
+--------------------------
+
+### Fixes
+
+- Fix an issue where scripts didn't put results on the document/console.
+- Fix an issue where the editor area was occasionally stacked under the window toolbar on macOS 10.12.
+- Fix an issue where MarsEdit via the App Store didn't update its contents after closing the document in CotEditor.
+- Improve general stability.
+
+
+
+3.0.4 (167)
 --------------------------
 
 ### Improvements